# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "DOCSWAP is a platform designed for realtors to efficiently share, manage, and analyze documents on a neighborhood level. It enables realtors to make better use of the content created for each listing by organizing it into a searchable database that supports both uploads and downloads of documents. These documents are categorized to facilitate market analysis, putting the power of data directly into the hands of realtors and market analysts interested in housing information in the lower mainland.",
        "title": "DOCSWAP API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "DOC-SWAP Support",
            "url": "http://www.doc-swap.com/support",
            "email": "support@doc-swap.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/tag/": {
            "get": {
                "description": "Get all tags from the database",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tags"
                ],
                "summary": "Retrieve all tags",
                "parameters": [
                    {
                        "type": "boolean",
                        "default": false,
                        "description": "Set to true to include soft deleted tags",
                        "name": "includeDeleted",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "default": false,
                        "description": "Set to true to include full tag details",
                        "name": "full",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved the tag",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Tag"
                            }
                        }
                    },
                    "400": {
                        "description": "Error: Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "description": "Create multiple new tag records",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tags"
                ],
                "summary": "Create multiple new tag records",
                "parameters": [
                    {
                        "description": "Tag objects",
                        "name": "tag",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Tag"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successfully created the tag records",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Error: Unable to create tag records",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/tag/bulk": {
            "delete": {
                "description": "Delete multiple tag records using tagID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tags"
                ],
                "summary": "Delete multiple tag records",
                "parameters": [
                    {
                        "description": "Tag objects that need to be deleted",
                        "name": "tags",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Tag"
                            }
                        }
                    },
                    {
                        "type": "boolean",
                        "default": true,
                        "description": "Set to false to fully delete the tag records",
                        "name": "softDelete",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successfully deleted the tag records",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Error: Unable to delete tag records",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/tag/search": {
            "post": {
                "description": "Search for tags with the given search parameters",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tags"
                ],
                "summary": "Search tags",
                "parameters": [
                    {
                        "description": "Search object",
                        "name": "searchObj",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/search.Search"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved the tag",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Tag"
                            }
                        }
                    },
                    "400": {
                        "description": "Error: Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/tag/{id}": {
            "get": {
                "description": "get tag record by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tags"
                ],
                "summary": "Retrieve tag record",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Tag ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "default": false,
                        "description": "Set to true to include soft deleted tags",
                        "name": "includeDeleted",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "default": false,
                        "description": "Set to true to include full tag details",
                        "name": "full",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved the tag",
                        "schema": {
                            "$ref": "#/definitions/models.Tag"
                        }
                    },
                    "400": {
                        "description": "Error: Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/tag/{tagId}": {
            "delete": {
                "description": "Delete a tag records via its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tags"
                ],
                "summary": "Delete a tag record",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Tag ID",
                        "name": "tagID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "default": true,
                        "description": "Set to false to fully delete the tag record",
                        "name": "softDelete",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successfully deleted the tag record",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Error: Failed to delete the tag record",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/user/": {
            "get": {
                "description": "get all user records",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Retrieve all users",
                "parameters": [
                    {
                        "type": "boolean",
                        "default": false,
                        "description": "Set to true to include soft deleted users",
                        "name": "includeDeleted",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "default": false,
                        "description": "Set to true to include full user details",
                        "name": "full",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved the list of users",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.User"
                            }
                        }
                    },
                    "400": {
                        "description": "Error: No user records found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "description": "create a new user record",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Create a new user",
                "parameters": [
                    {
                        "description": "User object that needs to be created",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successfully created the user",
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        }
                    },
                    "400": {
                        "description": "Error: Invalid request body",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.Category": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "parent": {
                    "$ref": "#/definitions/models.Category"
                },
                "parentID": {
                    "type": "integer"
                }
            }
        },
        "models.Config": {
            "type": "object",
            "properties": {
                "deletedAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
        "models.Document": {
            "type": "object",
            "properties": {
                "additionalDetails": {
                    "type": "string"
                },
                "address": {
                    "type": "string"
                },
                "approvedAt": {
                    "type": "string"
                },
                "approvedBy": {
                    "type": "integer"
                },
                "approvedByUser": {
                    "$ref": "#/definitions/models.User"
                },
                "category": {
                    "$ref": "#/definitions/models.Category"
                },
                "categoryID": {
                    "type": "integer"
                },
                "creditValue": {
                    "type": "integer"
                },
                "deletedAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "fileName": {
                    "type": "string"
                },
                "filePath": {
                    "type": "string"
                },
                "fileStorageURL": {
                    "type": "string"
                },
                "fileType": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.DocumentTag"
                    }
                },
                "uploadedAt": {
                    "type": "string"
                }
            }
        },
        "models.DocumentTag": {
            "type": "object",
            "properties": {
                "deletedAt": {
                    "type": "string"
                },
                "document": {
                    "$ref": "#/definitions/models.Document"
                },
                "documentID": {
                    "type": "integer"
                },
                "tag": {
                    "$ref": "#/definitions/models.Tag"
                },
                "tagID": {
                    "type": "integer"
                }
            }
        },
        "models.Permission": {
            "type": "object",
            "properties": {
                "deletedAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "roles": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.RolePermission"
                    }
                }
            }
        },
        "models.Role": {
            "type": "object",
            "properties": {
                "deletedAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "permissions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.RolePermission"
                    }
                },
                "users": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.UserRole"
                    }
                }
            }
        },
        "models.RolePermission": {
            "type": "object",
            "properties": {
                "deletedAt": {
                    "type": "string"
                },
                "permission": {
                    "$ref": "#/definitions/models.Permission"
                },
                "permissionID": {
                    "type": "integer"
                },
                "role": {
                    "$ref": "#/definitions/models.Role"
                },
                "roleID": {
                    "type": "integer"
                }
            }
        },
        "models.Tag": {
            "type": "object",
            "properties": {
                "deletedAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "documents": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.DocumentTag"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "models.User": {
            "type": "object",
            "properties": {
                "authenticationType": {
                    "type": "string"
                },
                "biography": {
                    "type": "string"
                },
                "configs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.UserConfig"
                    }
                },
                "deletedAt": {
                    "type": "string"
                },
                "documents": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.UserDocument"
                    }
                },
                "email": {
                    "type": "string"
                },
                "externalUserID": {
                    "type": "string"
                },
                "firstName": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "lastName": {
                    "type": "string"
                },
                "roles": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.UserRole"
                    }
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "models.UserConfig": {
            "type": "object",
            "properties": {
                "config": {
                    "$ref": "#/definitions/models.Config"
                },
                "configID": {
                    "type": "integer"
                },
                "deletedAt": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/models.User"
                },
                "userID": {
                    "type": "integer"
                }
            }
        },
        "models.UserDocument": {
            "type": "object",
            "properties": {
                "deletedAt": {
                    "type": "string"
                },
                "document": {
                    "$ref": "#/definitions/models.Document"
                },
                "documentID": {
                    "type": "integer"
                },
                "isOwner": {
                    "type": "boolean"
                },
                "user": {
                    "$ref": "#/definitions/models.User"
                },
                "userID": {
                    "type": "integer"
                }
            }
        },
        "models.UserRole": {
            "type": "object",
            "properties": {
                "deletedAt": {
                    "type": "string"
                },
                "role": {
                    "$ref": "#/definitions/models.Role"
                },
                "roleID": {
                    "type": "integer"
                },
                "user": {
                    "$ref": "#/definitions/models.User"
                },
                "userID": {
                    "type": "integer"
                }
            }
        },
...
        "search.LogicalOperator": {
            "type": "string",
            "enum": [
                "AND",
                "OR"
            ],
            "x-enum-varnames": [
                "And",
                "Or"
            ]
        },
        "search.Operator": {
            "type": "string",
            "enum": [
                "=",
                "!=",
                "\u003e",
                "\u003e=",
                "\u003c",
                "\u003c=",
                "LIKE",
                "NOT LIKE",
                "IN",
                "NOT IN",
                "IS NULL",
                "IS NOT NULL"
            ],
            "x-enum-varnames": [
                "Equal",
                "NotEqual",
                "GreaterThan",
                "GreaterThanOrEqual",
                "LessThan",
                "LessThanOrEqual",
                "Like",
                "NotLike",
                "In",
                "NotIn",
                "IsNull",
                "IsNotNull"
            ]
        },
        "search.Param": {
            "type": "object",
            "properties": {
                "AssociationForeignKey": {
                    "type": "string"
                },
                "Field": {
                    "type": "string"
                },
                "Operator": {
                    "$ref": "#/definitions/search.Operator"
                },
                "Value": {}
            }
        },
        "search.Search": {
            "type": "object",
            "properties": {
                "LogicalOperator": {
                    "$ref": "#/definitions/search.LogicalOperator"
                },
                "Params": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/search.Param"
                    }
                }
            }
        }
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    },
...
}